Change upload file page back to servercomponent
- Change made to allow page titling
- Imports changed to correct path
- Uploading not tested
