+++ src/app/legal/page.jsx
@@ -4,7 +4,7 @@ export default function LegalPage() {
             <p>We don&apos;t have a lawyer to write this out. Contact us immediately if any
                 copyrighted material is posted here. Otherwise, we try not to do
                 things we shouldn&apos;t, but should we severely misstep, do not hesitate to&nbsp;
-                <a href="mailto:legal@4nderground.com">contact us</a>.</p>
+                <a href="mailto:info@4nderground.com">contact us</a>.</p>
         </>
     );
 }
